fix(logic-function): rename migration to kebab-case + format
