+++ packages/twenty-server/src/database/commands/upgrade-version-command/2-8/2-8-instance-command-fast-1779534756998-add-logic-function-execution-mode.ts
