+++ packages/twenty-server/src/engine/metadata-modules/logic-function/dtos/logic-function.dto.ts
@@ -1,4 +1,9 @@
-import { Field, HideField, ObjectType, registerEnumType } from '@nestjs/graphql';
+import {
+  Field,
+  HideField,
+  ObjectType,
+  registerEnumType,
+} from '@nestjs/graphql';
 
 import {
   Authorize,
